+++ pkg/plugin/plugin.go
@@ -10,10 +10,7 @@ import (
 	"github.com/grafana/grafana-plugin-sdk-go/backend/instancemgmt"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
-	"github.com/grafana/grafana-plugin-sdk-go/live"
-	"math"
-	"regexp"
-	"strings"
+	"github.com/jmoiron/sqlx"
 	"time"
 )
 
@@ -95,100 +92,6 @@ type queryModel struct {
 	RawSqlSelected  bool   `json:"rawSqlSelected"`
 }
 
-func getIntervalString(duration time.Duration) string {
-	hours := int(math.Floor(duration.Hours()))
-	minutes := int(math.Floor(duration.Minutes()))
-	seconds := int(math.Floor(duration.Seconds()))
-
-	returnString := ""
-
-	deliminator := ""
-	if hours > 0 {
-		returnString = fmt.Sprintf("%s%s%d HOURS", returnString, deliminator, hours)
-		deliminator = " "
-	}
-
-	remainingMinutes := minutes - (hours * 60)
-	if remainingMinutes > 0 {
-		returnString = fmt.Sprintf("%s%s%d MINUTES", returnString, deliminator, remainingMinutes)
-		deliminator = " "
-	}
-
-	remainingSeconds := seconds - (minutes * 60)
-	if remainingSeconds > 0 {
-		returnString = fmt.Sprintf("%s%s%d SECONDS", returnString, deliminator, remainingSeconds)
-		deliminator = " "
-	}
-
-	return returnString
-}
-
-func replaceMacros(sqlQuery string, query backend.DataQuery) string {
-
-	queryString := sqlQuery
-	log.DefaultLogger.Info("Raw SQL Query selected", "query", queryString)
-
-	interval_string := getIntervalString(query.Interval)
-
-	var rgx = regexp.MustCompile(`\$__timeWindow\(([a-zA-Z0-9_-]+)\)`)
-	if rgx.MatchString(queryString) {
-		log.DefaultLogger.Info("__timeWindow placeholder found")
-		rs := rgx.FindStringSubmatch(queryString)
-		timeColumnName := rs[1]
-		queryString = rgx.ReplaceAllString(queryString, fmt.Sprintf("window(%s, '%s')", timeColumnName, interval_string))
-
-		rgx = regexp.MustCompile(`\$__time\(([a-zA-Z0-9_-]+)\)`)
-		if rgx.MatchString(queryString) {
-			log.DefaultLogger.Info("__time placeholder found")
-			queryString = rgx.ReplaceAllString(queryString, "window.start")
-		}
-
-		rgx = regexp.MustCompile(`\$__value\(([a-zA-Z0-9_-]+)\)`)
-		if rgx.MatchString(queryString) {
-			log.DefaultLogger.Info("__value placeholder found")
-			rs = rgx.FindStringSubmatch(queryString)
-			valueColumnName := rs[1]
-			queryString = rgx.ReplaceAllString(queryString, fmt.Sprintf("avg(%s) AS value", valueColumnName))
-		}
-	} else {
-		rgx = regexp.MustCompile(`\$__time\(([a-zA-Z0-9_-]+)\)`)
-		if rgx.MatchString(queryString) {
-			log.DefaultLogger.Info("__time placeholder found")
-			rs := rgx.FindStringSubmatch(queryString)
-			timeColumnName := rs[1]
-			queryString = rgx.ReplaceAllString(queryString, fmt.Sprintf("%s AS time", timeColumnName))
-		}
-
-		rgx = regexp.MustCompile(`\$__value\(([a-zA-Z0-9_-]+)\)`)
-		if rgx.MatchString(queryString) {
-			log.DefaultLogger.Info("__value placeholder found")
-			rs := rgx.FindStringSubmatch(queryString)
-			valueColumnName := rs[1]
-			queryString = rgx.ReplaceAllString(queryString, fmt.Sprintf("%s AS value", valueColumnName))
-		}
-	}
-
-	rgx = regexp.MustCompile(`\$__timeFilter\(([a-zA-Z0-9_-]+)\)`)
-	if rgx.MatchString(queryString) {
-		rs := rgx.FindStringSubmatch(queryString)
-		timeColumnName := rs[1]
-		timeRangeFilter := fmt.Sprintf("%s BETWEEN '%s' AND '%s'",
-			timeColumnName,
-			query.TimeRange.From.UTC().Format("2006-01-02 15:04:05"),
-			query.TimeRange.To.UTC().Format("2006-01-02 15:04:05"),
-		)
-		queryString = rgx.ReplaceAllString(queryString, timeRangeFilter)
-	}
-
-	queryString = strings.ReplaceAll(queryString, "$__timeFrom", query.TimeRange.From.UTC().Format("2006-01-02 15:04:05"))
-
-	queryString = strings.ReplaceAll(queryString, "$__timeTo", query.TimeRange.To.UTC().Format("2006-01-02 15:04:05"))
-
-	queryString = strings.ReplaceAll(queryString, "$__interval", interval_string)
-
-	return queryString
-}
-
 func (d *SampleDatasource) query(_ context.Context, pCtx backend.PluginContext, query backend.DataQuery) backend.DataResponse {
 	response := backend.DataResponse{}
 
@@ -203,109 +106,53 @@ func (d *SampleDatasource) query(_ context.Context, pCtx backend.PluginContext, 
 		return response
 	}
 
-	seconds_interval := int64(query.Interval / 1000000000)
-	log.DefaultLogger.Info("Querry Interval (sec):", seconds_interval)
-	if seconds_interval <= 0 {
-		seconds_interval = 1
-	}
-	interval_string := getIntervalString(query.Interval)
 	queryString := ""
 	if qm.RawSqlSelected {
 		queryString = replaceMacros(qm.RawSqlQuery, query)
 	} else {
-		whereQuery := ""
-		if qm.WhereQuery != "" {
-			whereQuery = fmt.Sprintf(" %s AND", qm.WhereQuery)
-		}
-		queryString = fmt.Sprintf("SELECT window.start, avg(%s) AS value FROM %s WHERE%s %s BETWEEN '%s' AND '%s' GROUP BY window(%s, '%s')",
-			qm.ValueColumnName,
-			qm.TableName,
-			whereQuery,
-			qm.TimeColumnName,
-			query.TimeRange.From.UTC().Format("2006-01-02 15:04:05"),
-			query.TimeRange.To.UTC().Format("2006-01-02 15:04:05"),
-			qm.TimeColumnName,
-			interval_string)
+		queryString = createQueryBuilderQuery(query, qm)
 	}
 	log.DefaultLogger.Info("Query", "query", queryString)
 
-	rows, err := databricksDB.Query(queryString)
-	defer rows.Close()
+	frame := data.NewFrame("response")
 
+	db := sqlx.NewDb(databricksDB, "postgres")
+
+	rows, err := db.Queryx(queryString)
 	if err != nil {
 		response.Error = err
-		log.DefaultLogger.Info("Query Execution Error", "err", err)
+		log.DefaultLogger.Info("Error", "err", err)
 		return response
 	}
 
-	rowCount := 0
-	for rows.Next() {
-		rowCount = rowCount + 1
-	}
-
-	rows, _ = databricksDB.Query(queryString)
-	cols, err := rows.Columns() // Remember to check err afterwards
+	colTypes, err := rows.ColumnTypes()
 
 	if err != nil {
 		response.Error = err
-		log.DefaultLogger.Info("Extracting Columns Error", "err", err)
+		log.DefaultLogger.Info("Error", "err", err)
+		return response
+	}
+	columnNames, err := rows.Columns()
+	if err != nil {
+		response.Error = err
+		log.DefaultLogger.Info("Error", "err", err)
 		return response
 	}
 
-	log.DefaultLogger.Info("Columns:", cols)
-	vals := make([]interface{}, len(cols))
-
-	timestamps := make([]time.Time, rowCount)
-	values := make([]float32, rowCount)
-
-	for i, _ := range cols {
-		vals[i] = new(sql.RawBytes)
+	frame, err = initDataframeTypes(colTypes, columnNames, frame)
+	if err != nil {
+		response.Error = err
+		return response
 	}
-	i := 0
+
 	for rows.Next() {
-		var (
-			timestamp time.Time
-			value     float32
-		)
-		err = rows.Scan(&timestamp, &value)
+		res, err := rows.SliceScan()
 		if err != nil {
 			response.Error = err
-			log.DefaultLogger.Info("Row Scan Error", "err", err)
+			log.DefaultLogger.Info("Error", "err", err)
 			return response
 		}
-		//log.DefaultLogger.Info("Row Next", "err", err)
-		//log.DefaultLogger.Info("Returned timestamp", timestamp)
-		//log.DefaultLogger.Info("Returned value", value)
-		timestamps[i] = timestamp
-		values[i] = value
-		i = i + 1
-	}
-	err = rows.Err()
-
-	if err != nil {
-		response.Error = err
-		log.DefaultLogger.Info("Row Error", "err", err)
-		return response
-	}
-
-	// create data frame response.
-	frame := data.NewFrame("response")
-	frame.Fields = append(frame.Fields,
-		data.NewField("timestamp", nil, timestamps),
-		data.NewField(qm.ValueColumnName, nil, values),
-	)
-	// add fields.
-
-	// If query called with streaming on then return a channel
-	// to subscribe on a client-side and consume updates from a plugin.
-	// Feel free to remove this if you don't need streaming for your datasource.
-	if qm.WithStreaming {
-		channel := live.Channel{
-			Scope:     live.ScopeDatasource,
-			Namespace: pCtx.DataSourceInstanceSettings.UID,
-			Path:      "stream",
-		}
-		frame.SetMeta(&data.FrameMeta{Channel: channel.String()})
+		frame.AppendRow(res...)
 	}
 
 	// add the frames to the response.
@@ -329,15 +176,6 @@ func (d *SampleDatasource) CheckHealth(_ context.Context, req *backend.CheckHeal
 			Message: "No connection string found." + "Set the DATABRICKS_DSN environment variable, and try again.",
 		}, nil
 	}
-
-	//db, err := sql.Open("databricks", dsn)
-	//
-	//if err != nil {
-	//	return &backend.CheckHealthResult{
-	//		Status:  backend.HealthStatusError,
-	//		Message: fmt.Sprintf("SQL Connection Failed: %s", err),
-	//	}, nil
-	//}
 
 	rows, err := databricksDB.Query("SELECT 1")