+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
-## 0.0.5 (Unreleased)
+## 0.0.6
 
-Initial release.
+- Adds support for multiple metrics queries (SQL Editor only)
+- Fixes bug where saved query was not displayed in query builder
+
+---
+
+### 0.0.5
+
+- Adds full text SQL query editor with macros